Struct lexical_write_float::OptionsBuilder

pub struct OptionsBuilder { /* fields omitted */ }

Builder for Options.

Implementations

impl OptionsBuilder

pub const fn new() -> Self

pub const fn get_max_significant_digits(&self) -> OptionUsize

Get the maximum number of significant digits to write.

pub const fn get_min_significant_digits(&self) -> OptionUsize

Get the minimum number of significant digits to write.

pub const fn get_positive_exponent_break(&self) -> OptionI32

Get the maximum exponent prior to using scientific notation.

pub const fn get_negative_exponent_break(&self) -> OptionI32

Get the minimum exponent prior to using scientific notation.

pub const fn get_round_mode(&self) -> RoundMode

Get the rounding mode for writing digits with precision control.

pub const fn get_trim_floats(&self) -> bool

Get if we should trim a trailing ".0" from floats.

pub const fn get_exponent(&self) -> u8

Get the character to designate the exponent component of a float.

pub const fn get_decimal_point(&self) -> u8

Get the character to separate the integer from the fraction components.

pub const fn get_nan_string(&self) -> Option<&'static [u8]>

Get the string representation for NaN.

pub const fn get_inf_string(&self) -> Option<&'static [u8]>

Get the short string representation for Infinity.

pub const fn max_significant_digits(
self,
max_significant_digits: OptionUsize
) -> Self

Set the maximum number of significant digits to write.

pub const fn min_significant_digits(
self,
min_significant_digits: OptionUsize
) -> Self

Set the minimum number of significant digits to write.

pub const fn positive_exponent_break(
self,
positive_exponent_break: OptionI32
) -> Self

Set the maximum exponent prior to using scientific notation.

pub const fn negative_exponent_break(
self,
negative_exponent_break: OptionI32
) -> Self

Set the minimum exponent prior to using scientific notation.

pub const fn round_mode(self, round_mode: RoundMode) -> Self

Set the rounding mode for writing digits with precision control.

pub const fn trim_floats(self, trim_floats: bool) -> Self

Set if we should trim a trailing ".0" from floats.

pub const fn exponent(self, exponent: u8) -> Self

Set the character to designate the exponent component of a float.

pub const fn decimal_point(self, decimal_point: u8) -> Self

Set the character to separate the integer from the fraction components.

pub const fn nan_string(self, nan_string: Option<&'static [u8]>) -> Self

Set the string representation for NaN.

pub const fn inf_string(self, inf_string: Option<&'static [u8]>) -> Self

Set the string representation for Infinity.

pub const fn nan_str_is_valid(&self) -> bool

Determine if nan_str is valid.

pub const fn inf_str_is_valid(&self) -> bool

Determine if inf_str is valid.

pub const fn is_valid(&self) -> bool

Check if the builder state is valid.

pub const unsafe fn build_unchecked(&self) -> Options

Build the Options struct with bounds validation.

Safety

Safe as long as is_valid is true. If nan_string or inf_string are too long, writing special floats may lead to buffer overflows, and therefore severe security vulnerabilities.

pub const fn build(&self) -> Result<Options>

Build the Options struct.